Move things to a source folder.
